BIM: Fix Arch_MergeWalls output dependency on view orientation
Currently if user uses Arch_MergeWalls with walls based on Line objects,
the result depends on the view orientation (while working with Top
orientation correctly).

The problem is because of `makeSketch` using the current view direction
to determine the sketch's orientation when creating a new sketch from
Line objects which do not define a plane. This caused walls to be
correctly merged in top view but rotated in other views.

The fix is to preserve the correct placement of the base wall and apply
it to new sketch upon creation, so it won't get messed up in the
process.